+++ eno-core/src/main/java/fr/insee/eno/core/model/calculated/CalculatedExpression.java
@@ -8,11 +8,13 @@ import fr.insee.eno.core.exceptions.business.IllegalPoguesElementException;
 import fr.insee.eno.core.model.EnoObject;
 import fr.insee.eno.core.parameter.Format;
 import fr.insee.eno.core.reference.PoguesIndex;
+import fr.insee.eno.core.reference.VariableIndex;
 import fr.insee.lunatic.model.flat.LabelType;
 import fr.insee.lunatic.model.flat.LabelTypeEnum;
 import fr.insee.pogues.model.ExpressionType;
 import lombok.Getter;
 import lombok.Setter;
+import lombok.extern.slf4j.Slf4j;
 
 import java.util.ArrayList;
 import java.util.LinkedHashSet;
@@ -27,6 +29,7 @@ import static fr.insee.eno.core.annotations.Contexts.Context;
  * calculated variables, controls, filters. */
 @Getter
 @Setter
+@Slf4j
 @Context(format = Format.POGUES, type = ExpressionType.class)
 @Context(format = Format.DDI, type = CommandType.class)
 @Context(format = Format.LUNATIC, type = LabelType.class)
@@ -69,25 +72,28 @@ public class CalculatedExpression extends EnoObject {
     /** The method "extractBindingReferences" allows, from the expression of a "CalculatedVariable",
      * to construct the associated "BindingReferences". Note that the identifier (a concept derived from DDI)
      * is null for each reference. */
-    public static Set<BindingReference> extractBindingReferences(String expression, PoguesIndex poguesIndex) {
+    public static Set<BindingReference> extractBindingReferences(String expression, VariableIndex variableIndex) {
         Set<BindingReference> references = new LinkedHashSet<>(); // linked hash set to have consistent order
         Matcher matcher = POGUES_VARIABLE_PATTERN.matcher(expression);
 
         while (matcher.find()) {
             String variableName = matcher.group(1);
-            validatePoguesReference(expression, variableName, poguesIndex);
+            validatePoguesReference(expression, variableName, variableIndex);
             references.add(new BindingReference(null, variableName));
         }
 
         return references;
     }
 
-    private static void validatePoguesReference(String expression, String variableName, PoguesIndex poguesIndex) {
-        if (! poguesIndex.containsVariable(variableName))
-            throw new IllegalPoguesElementException(String.format(
+    private static void validatePoguesReference(String expression, String variableName, VariableIndex variableIndex) {
+        if (!variableIndex.containsVariable(variableName)) {
+            String message = String.format(
                     "Name '%s' used in expression:%n%s%n" +
                             "does not match any variable.",
-                    variableName, expression));
+                    variableName, expression);
+            log.warn(message);
+            // should be an exception, yet Pogues composition feature creates cases where this is allowed
+        }
     }
 
     /** The removeSurroundingDollarSigns method removes the "$" symbols surrounding the reference to