+++ eno-core/src/main/java/fr/insee/eno/core/model/question/DateQuestion.java
@@ -59,9 +59,10 @@ public class DateQuestion extends SingleResponseQuestion {
     @Lunatic("setDateFormat(#param)")
     private String format;
 
-    /** Lunatic component type property.
-     * This should be inserted by Lunatic-Model serializer later on. */
-    @Lunatic("setComponentType(T(fr.insee.lunatic.model.flat.ComponentTypeEnum).valueOf(#param))")
-    String lunaticComponentType = "DATEPICKER";
+    /** Indicates whether the response is mandatory for this component. */
+    @DDI("getResponseDomain()?.getResponseCardinality()?.getMinimumResponses() != null ? " +
+            "getResponseDomain().getResponseCardinality().getMinimumResponses().intValue() > 0 : false")
+    @Lunatic("setMandatory(#param)")
+    boolean mandatory;
 
 }