+++ eno-core/src/main/java/fr/insee/eno/core/model/question/UniqueChoiceQuestion.java
@@ -11,6 +11,7 @@ import fr.insee.eno.core.exceptions.technical.ConversionException;
 import fr.insee.eno.core.exceptions.technical.MappingException;
 import fr.insee.eno.core.model.code.CodeItem;
 import fr.insee.eno.core.model.navigation.Binding;
+import fr.insee.eno.core.model.response.CodeFilter;
 import fr.insee.eno.core.model.response.DetailResponse;
 import fr.insee.eno.core.model.response.Response;
 import fr.insee.eno.core.parameter.Format;
@@ -99,6 +100,18 @@ public class UniqueChoiceQuestion extends SingleResponseQuestion {
     @Pogues("getClarificationQuestion()")
     @DDI("T(fr.insee.eno.core.model.question.UniqueChoiceQuestion).mapDetailResponses(#this)")
     List<DetailResponse> detailResponses = new ArrayList<>();
+
+    /**
+     * List of conditions for the modalities to be filtered by previous responses or external data.
+     * In Lunatic, they are inserted in option through a processing. */
+    @Pogues("getCodeFilters()")
+    List<CodeFilter> codeFilters = new ArrayList<>();
+
+    /** Indicates whether the response is mandatory for this component. */
+    @DDI("getResponseDomain()?.getResponseCardinality()?.getMinimumResponses() != null ? " +
+            "getResponseDomain().getResponseCardinality().getMinimumResponses().intValue() > 0 : false")
+    @Lunatic("setMandatory(#param)")
+    boolean mandatory;
 
     public static DisplayFormat convertPoguesVisualizationHint(ResponseType poguesResponse) {
         return switch (poguesResponse.getDatatype().getVisualizationHint()) {