+++ eno-core/src/main/java/fr/insee/eno/core/model/response/CodeResponse.java
@@ -43,4 +43,9 @@ public class CodeResponse extends EnoIdentifiableObject {
     @Lunatic("setDetail(#param)")
     DetailResponse detailResponse;
 
+    /** Condition for filtering the modality.
+     * In Pogues, it is mapped at the question level and inserted here in a processing step. */
+    @Lunatic("setConditionFilter(#param)")
+    CodeFilter conditionFilter;
+
 }