+++ eno-core/src/test/java/fr/insee/eno/core/processing/out/steps/lunatic/LunaticAddCleaningTest.java
@@ -2,7 +2,9 @@ package fr.insee.eno.core.processing.out.steps.lunatic;
 
 import fr.insee.ddi.lifecycle33.instance.DDIInstanceDocument;
 import fr.insee.eno.core.DDIToEno;
+import fr.insee.eno.core.PoguesDDIToEno;
 import fr.insee.eno.core.exceptions.business.DDIParsingException;
+import fr.insee.eno.core.exceptions.business.PoguesDeserializationException;
 import fr.insee.eno.core.mappers.LunaticMapper;
 import fr.insee.eno.core.model.EnoQuestionnaire;
 import fr.insee.eno.core.model.navigation.Filter;
@@ -11,11 +13,15 @@ import fr.insee.eno.core.parameter.EnoParameters;
 import fr.insee.eno.core.parameter.Format;
 import fr.insee.eno.core.parameter.LunaticParameters;
 import fr.insee.eno.core.processing.ProcessingPipeline;
+import fr.insee.eno.core.processing.out.steps.lunatic.cleaning.CleaningUtils;
+import fr.insee.eno.core.processing.out.steps.lunatic.cleaning.LunaticAddCleaning;
 import fr.insee.eno.core.processing.out.steps.lunatic.pagination.LunaticAddPageNumbers;
 import fr.insee.eno.core.processing.out.steps.lunatic.resizing.LunaticAddResizing;
 import fr.insee.eno.core.processing.out.steps.lunatic.table.LunaticTableProcessing;
 import fr.insee.eno.core.reference.EnoIndex;
 import fr.insee.eno.core.serialize.DDIDeserializer;
+import fr.insee.eno.core.serialize.LunaticSerializer;
+import fr.insee.eno.core.serialize.PoguesDeserializer;
 import fr.insee.lunatic.model.flat.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -23,12 +29,11 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.*;
 
-class LunaticAddCleaningVariablesTest {
+class LunaticAddCleaningTest {
 
-    LunaticAddCleaningVariables cleaningProcessing;
+    LunaticAddCleaning cleaningProcessing;
     Questionnaire lunaticQuestionnaire;
     EnoQuestionnaire enoQuestionnaire;
 
@@ -69,8 +74,9 @@ class LunaticAddCleaningVariablesTest {
     void testGetFinalBindingReferences() throws DDIParsingException {
         prepareQuestionnaireTest("functional/ddi/cleaning/ddi-m7oqvx8y.xml");
         Filter calculatedFilter = (Filter) enoQuestionnaire.getIndex().get("m7oqxnxe");
-        List<String> finalBindingsDeps = cleaningProcessing.removeCalculatedVariables(
-                cleaningProcessing.getFinalBindingReferencesWithCalculatedVariables(calculatedFilter.getExpression()));
+        List<String> finalBindingsDeps = CleaningUtils.removeCalculatedVariables(
+                CleaningUtils.getFinalBindingReferencesWithCalculatedVariables(calculatedFilter.getExpression(), cleaningProcessing.getVariableIndex()),
+                cleaningProcessing.getVariableIndex());
         assertThat(finalBindingsDeps).containsExactly("AGE");
 
     }
@@ -84,16 +90,75 @@ class LunaticAddCleaningVariablesTest {
         assertNull(cleaningProcessing.getVariableShapeFromIndex().get(variableOutOfLoop));
     }
 
+    @Test
+    void testCleaningOfCodeFilters() throws DDIParsingException, PoguesDeserializationException {
+        prepareQuestionnaireTest(
+                "functional/pogues/codes-filtered/pogues-m8hgkyw0.json",
+                "functional/ddi/codes-filtered/ddi-m8hgkyw0.xml");
+        cleaningProcessing.processCodeFilters(lunaticQuestionnaire);
+        assertNotNull(lunaticQuestionnaire.getCleaning().getCleaningEntry("AGE"));
+        assertThat(lunaticQuestionnaire.getCleaning()
+                .getCleaningEntry("AGE")
+                .getCleanedVariable("RADIO_OUI_NON")
+                .getCleaningExpressions())
+                .hasSize(2);
+        assertEquals("(nvl(AGE, 0) > 18) or (RADIO_OUI_NON <> \"3\")", lunaticQuestionnaire.getCleaning()
+                .getCleaningEntry("AGE")
+                .getCleanedVariable("RADIO_OUI_NON")
+                .getCleaningExpressions().get(0).getExpression());
+        assertEquals("(nvl(AGE, 0) > 50) or (RADIO_OUI_NON <> \"4\")", lunaticQuestionnaire.getCleaning()
+                .getCleaningEntry("AGE")
+                .getCleanedVariable("RADIO_OUI_NON")
+                .getCleaningExpressions().get(1).getExpression());
 
-    void prepareQuestionnaireTest(String questionnaireTestUrl) throws DDIParsingException {
+        assertThat(lunaticQuestionnaire.getCleaning()
+                .getCleaningEntry("AGE")
+                .getCleanedVariable("DROPDOWN_OUI_NON")
+                .getCleaningExpressions())
+                .hasSize(1);
+        assertEquals("(nvl(AGE, 0) > 18) or (DROPDOWN_OUI_NON <> \"3\")", lunaticQuestionnaire.getCleaning()
+                .getCleaningEntry("AGE")
+                .getCleanedVariable("DROPDOWN_OUI_NON")
+                .getCleaningExpressions().get(0).getExpression());
+
+        assertThat(lunaticQuestionnaire.getCleaning()
+                .getCleaningEntry("AGE")
+                .getCleanedVariable("CHEXBOXMULTI_OUI_NON3")
+                .getCleaningExpressions())
+                .hasSize(1);
+        assertEquals("nvl(AGE, 0) > 18", lunaticQuestionnaire.getCleaning()
+                .getCleaningEntry("AGE")
+                .getCleanedVariable("CHEXBOXMULTI_OUI_NON3")
+                .getCleaningExpressions().get(0).getExpression());
+
+        System.out.println(LunaticSerializer.serializeToJson(lunaticQuestionnaire));
+    }
+
+    void prepareQuestionnaireTest(String poguesQuestionnaireTestUrl, String ddiQuestionnaireTestUrl) throws DDIParsingException, PoguesDeserializationException {
         EnoParameters enoParameters = EnoParameters.of(EnoParameters.Context.DEFAULT, EnoParameters.ModeParameter.CAWI, Format.LUNATIC);
+        fr.insee.pogues.model.Questionnaire poguesQuestionnaire = PoguesDeserializer.deserialize(
+                this.getClass().getClassLoader().getResourceAsStream(poguesQuestionnaireTestUrl));
         DDIInstanceDocument ddiQuestionnaire = DDIDeserializer.deserialize(
-                this.getClass().getClassLoader().getResourceAsStream(questionnaireTestUrl));
-        enoQuestionnaire = DDIToEno.fromObject(ddiQuestionnaire).transform(enoParameters);
-        cleaningProcessing = new LunaticAddCleaningVariables(enoQuestionnaire);
+                this.getClass().getClassLoader().getResourceAsStream(ddiQuestionnaireTestUrl));
+        enoQuestionnaire = PoguesDDIToEno.fromObjects(poguesQuestionnaire, ddiQuestionnaire).transform(enoParameters);
+        cleaningProcessing = new LunaticAddCleaning(enoQuestionnaire);
         lunaticQuestionnaire = new Questionnaire();
         preProcessQuestionnaire(lunaticQuestionnaire, enoQuestionnaire);
         applyProcessingBeforeCleaning(lunaticQuestionnaire, enoQuestionnaire);
+        cleaningProcessing.preProcessCleaning(lunaticQuestionnaire);
+        cleaningProcessing.preProcessVariablesAndShapeFrom(lunaticQuestionnaire);
+    }
+
+    void prepareQuestionnaireTest(String ddiQuestionnaireTestUrl) throws DDIParsingException {
+        EnoParameters enoParameters = EnoParameters.of(EnoParameters.Context.DEFAULT, EnoParameters.ModeParameter.CAWI, Format.LUNATIC);
+        DDIInstanceDocument ddiQuestionnaire = DDIDeserializer.deserialize(
+                this.getClass().getClassLoader().getResourceAsStream(ddiQuestionnaireTestUrl));
+        enoQuestionnaire = DDIToEno.fromObject(ddiQuestionnaire).transform(enoParameters);
+        cleaningProcessing = new LunaticAddCleaning(enoQuestionnaire);
+        lunaticQuestionnaire = new Questionnaire();
+        preProcessQuestionnaire(lunaticQuestionnaire, enoQuestionnaire);
+        applyProcessingBeforeCleaning(lunaticQuestionnaire, enoQuestionnaire);
+        cleaningProcessing.preProcessCleaning(lunaticQuestionnaire);
         cleaningProcessing.preProcessVariablesAndShapeFrom(lunaticQuestionnaire);
     }
 
@@ -112,6 +177,7 @@ class LunaticAddCleaningVariablesTest {
                 .then(new LunaticLoopResolution(enoQuestionnaire))
                 .then(new LunaticTableProcessing(enoQuestionnaire))
                 .then(new LunaticInsertUniqueChoiceDetails(enoQuestionnaire))
+                .then(new LunaticInsertCodeFilters(enoQuestionnaire))
                 .then(new LunaticEditLabelTypes()) // this step should be temporary
                 .then(new LunaticSuggestersConfiguration(enoQuestionnaire))
                 .then(new LunaticVariablesDimension(enoQuestionnaire))