+++ eno-core/src/test/java/fr/insee/eno/core/utils/vtl/VtlSyntaxUtilsTest.java
@@ -27,6 +27,7 @@ class VtlSyntaxUtilsTest {
         assertEquals(")", VtlSyntaxUtils.getVTLTokenName(VtlTokens.RPAREN));
         assertEquals(";", VtlSyntaxUtils.getVTLTokenName(VtlTokens.EOL));
         assertEquals(":=", VtlSyntaxUtils.getVTLTokenName(VtlTokens.ASSIGN));
+        assertEquals("<>", VtlSyntaxUtils.getVTLTokenName(VtlTokens.NEQ));
     }
 
     @Test
@@ -37,6 +38,37 @@ class VtlSyntaxUtilsTest {
         assertFalse(VtlSyntaxUtils.isAggregatorUsedInsideExpression(expressionWithoutAgg));
         String trickyExpressionWithoutAgg = "\"count(FIRST_NAME)\" || FIRST_NAME"; // here 'count' is not an operator since it is in a string
         assertFalse(VtlSyntaxUtils.isAggregatorUsedInsideExpression(trickyExpressionWithoutAgg));
+    }
+
+    @Test
+    void testSurroundByParenthesis(){
+        assertEquals("(FIRST_NAME = \"Laurent\")", VtlSyntaxUtils.surroundByParenthesis("FIRST_NAME = \"Laurent\""));
+        assertEquals("((NB >= 15))", VtlSyntaxUtils.surroundByParenthesis("(NB >= 15)"));
+    }
+
+    @Test
+    void testSurroundByDoubleQuotes(){
+        assertEquals("\"FIRST_NAME\"", VtlSyntaxUtils.surroundByDoubleQuotes("FIRST_NAME"));
+    }
+
+    @Test
+    void testJoinByORLogicExpression(){
+        String expectedExpression = "(FIRST_NAME = \"Laurent\") or (NB >= 15)";
+        assertEquals(expectedExpression, VtlSyntaxUtils.joinByORLogicExpression("FIRST_NAME = \"Laurent\"", "NB >= 15"));
+    }
+
+    @Test
+    void testJoinByANDLogicExpression_should_addAndVtlBetweenExpression(){
+        String expectedExpression = "(FIRST_NAME = \"Laurent\") and (NB >= 15)";
+        assertEquals(expectedExpression, VtlSyntaxUtils.joinByANDLogicExpression("FIRST_NAME = \"Laurent\"", "NB >= 15"));
+    }
+
+    @Test
+    void testJoinByANDLogicExpression_should_not_addExtraParen_when_AddAndVtlBetweenExpression(){
+        String expectedExpression = "(FIRST_NAME = \"Laurent\") and (NB >= 15)";
+        assertEquals(expectedExpression, VtlSyntaxUtils.joinByANDLogicExpression(
+                VtlSyntaxUtils.surroundByParenthesis("FIRST_NAME = \"Laurent\""),
+                VtlSyntaxUtils.surroundByParenthesis("NB >= 15")));
     }
 
 }